Move watcher from cmd/manager/internal/agentmap to mutator
