+++ cmd/traffic/cmd/manager/internal/mutator/service.go
@@ -23,7 +23,6 @@ import (
 	"github.com/datawire/dlib/dhttp"
 	"github.com/datawire/dlib/dlog"
 	"github.com/datawire/dlib/dtime"
-	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/internal/agentmap"
 	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/managerutil"
 	"github.com/telepresenceio/telepresence/v2/pkg/install"
 )
@@ -118,7 +117,7 @@ func ServeMutator(ctx context.Context) error {
 	})
 
 	env := managerutil.GetEnv(ctx)
-	cw, err := agentmap.Load(ctx, env.ManagerNamespace)
+	cw, err := Load(ctx, env.ManagerNamespace)
 	if err != nil {
 		return err
 	}
